+++ spec/timezone.html
@@ -163,7 +163,8 @@
         The following steps are taken:
       </p>
       <emu-alg>
-        1. Return ? TimeZoneFrom(_item_).
+        1. Let _constructor_ be the *this* value.
+        1. Return ? TimeZoneFrom(_item_, _constructor_).
       </emu-alg>
     </emu-clause>
   </emu-clause>
@@ -561,12 +562,12 @@
       <emu-alg>
         1. If Type(_temporalTimeZoneLike_) is Object, return _temporalTimeZoneLike_.
         1. Let _identifier_ be ? ToString(_temporalTimeZoneLike_).
-        1. Return ? TimeZoneFrom(_identifier_).
+        1. Return ? TimeZoneFrom(_identifier_, %Temporal.TimeZone%).
       </emu-alg>
     </emu-clause>
 
     <emu-clause id="sec-temporal-timezonefrom" aoid="TimeZoneFrom">
-      <h1>TimeZoneFrom ( _identifier_ )</h1>
+      <h1>TimeZoneFrom ( _identifier_, _constructor_ )</h1>
       <emu-alg>
         1. If Type(_item_) is Object, then
           1. If ? HasProperty(_item_, *"timeZone"*) is *false*, return _item_.
@@ -574,7 +575,6 @@
           1. If Type(_item_) is Object and ? HasProperty(_item_, *"timeZone"*) is *false*, return _item_.
         1. Let _string_ be ? ToString(_item_).
         1. Let _result_ be ? ParseTemporalTimeZone(_string_).
-        1. Let _constructor_ be the *this* value.
         1. Return ? CreateTemporalTimeZoneFromStatic(_constructor_, _result_).
       </emu-alg>
     </emu-clause>
