+++ spec/zoneddatetime.html
@@ -1139,7 +1139,7 @@
           1. Let _string_ be ? ToString(_item_).
           1. Let _result_ be ? ParseTemporalZonedDateTimeString(_string_).
           1. If _result_.[[TimeZoneName]] is *undefined*, throw a *RangeError* exception.
-          1. Let _timeZone_ be ? TimeZoneFrom(_result_.[[TimeZoneName]]).
+          1. Let _timeZone_ be ? TimeZoneFrom(_result_.[[TimeZoneName]], %Temporal.TimeZone%).
           1. Let _offsetString_ be _result_.[[TimeZoneOffsetString]].
           1. Let _calendar_ be ? ToOptionalTemporalCalendar(_result_.[[Calendar]]).
         1. Let _offsetNanoseconds_ be ? ParseTimeZoneOffsetString(_offsetString_).
